Update tests to support older verions of Rollbar.

+++ tests/TestCase.php
@@ -3,6 +3,7 @@
 namespace ShiftOneLabs\PhpRollbarExtensions\Tests;
 
 use Rollbar\Rollbar;
+use ReflectionProperty;
 use PHPUnit_Framework_TestCase;
 
 class TestCase extends PHPUnit_Framework_TestCase
@@ -13,12 +14,26 @@ class TestCase extends PHPUnit_Framework_TestCase
 
     public function tearDown()
     {
-        Rollbar::destroy();
+        $this->clearLogger();
+
         parent::tearDown();
     }
 
     public function getTestAccessToken()
     {
         return isset($_ENV['ROLLBAR_TEST_TOKEN']) ? $_ENV['ROLLBAR_TEST_TOKEN'] : static::DEFAULT_ACCESS_TOKEN;
+    }
+
+    protected function clearLogger()
+    {
+        // Rollbar::destroy() wasn't introduced until 1.4.1.
+        if (method_exists(Rollbar::class, 'destroy')) {
+            return Rollbar::destroy();
+        }
+
+        // Use reflection to clear the logger for 1.2.0 - 1.4.0.
+        $loggerProperty = new ReflectionProperty(Rollbar::class, 'logger');
+        $loggerProperty->setAccessible(true);
+        $loggerProperty->setValue(null);
     }
 }
